fix(hotkeys): label ctrl bindings as Cmd on macOS
The Settings tab rendered a `ctrl` binding as "Ctrl" on every platform, but
hotkeyStore resolves that same binding to the platform primary modifier — Meta
on macOS. Undo therefore advertised Ctrl+Z while only Cmd+Z actually fired.

The label helpers never consulted the platform at all, so move them out of the
component into hotkeyLabels.ts, where they take the primary modifier label as an
argument and can be unit-tested without React. The tab feeds them
usePlatformModifier(), which resolves after mount and so keeps SSR output stable.

Reordering toKeyLabel's checks along the way: Space is one character long, so
the upper-case branch used to swallow it and render a blank keycap.
